# src/main/resources/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hệ thống quản lý tài sản và trang thiết bị của cán bộ",
    "description": "API chia sẻ khai thác thông tin tài sản và hồ sơ cán bộ chiến sĩ qua LGSP.",
    "contact": {
      "name": "Ban quản lý hệ thống"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/chia-se/api",
      "description": "Local Dev"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/v1/can-bo": {
      "get": {
        "tags": [
          "1. Dịch vụ khai thác dữ liệu thông tin cán bộ chiến sĩ qua LGSP"
        ],
        "summary": "Lấy danh sách cán bộ chiến sĩ",
        "operationId": "getCanBo",
        "parameters": [
          {
            "name": "maDonVi",
            "in": "query",
            "description": "Mã đơn vị",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maCanBo",
            "in": "query",
            "description": "Mã cán bộ",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hoTen",
            "in": "query",
            "description": "Họ và tên (tìm kiếm gần đúng)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "soCccd",
            "in": "query",
            "description": "Số CCCD",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "capBac",
            "in": "query",
            "description": "Cấp bậc",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Số thứ tự trang (bắt đầu từ 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số lượng bản ghi trên một trang",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy dữ liệu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponsePagingBaseResponseCanBoTomTatResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "description": "API tra cứu danh sách cán bộ chiến sĩ, phục vụ việc chia sẻ khai thác thông tin (có hỗ trợ phân trang)."
      }
    },
    "/v1/can-bo/{id}": {
      "get": {
        "tags": [
          "1. Dịch vụ khai thác dữ liệu thông tin cán bộ chiến sĩ qua LGSP"
        ],
        "summary": "Lấy thông tin chi tiết một cán bộ chiến sĩ",
        "operationId": "getCanBoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy dữ liệu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponseCanBoChiTietResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "description": "API lấy thông tin chi tiết của một cán bộ chiến sĩ, bao gồm thông tin tài sản được giao."
      }
    },
    "/v1/tai-san": {
      "get": {
        "tags": [
          "2. Dịch vụ khai thác dữ liệu thông tin tài sản qua LGSP"
        ],
        "summary": "Lấy danh sách tài sản",
        "operationId": "getTaiSan",
        "parameters": [
          {
            "name": "maDonViQuanLy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maCanBoSuDung",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tinhTrang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Số thứ tự trang (bắt đầu từ 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số lượng bản ghi trên một trang",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy dữ liệu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponsePagingBaseResponseTaiSanDtoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "description": "API tra cứu danh sách tài sản."
      }
    },
    "/v1/tai-san/{id}": {
      "get": {
        "tags": [
          "2. Dịch vụ khai thác dữ liệu thông tin tài sản qua LGSP"
        ],
        "summary": "Lấy chi tiết tài sản",
        "operationId": "getTaiSanById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy dữ liệu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponseTaiSanDtoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "description": "API lấy thông tin chi tiết của một tài sản."
      }
    },
    "/v1/tai-san/{id}/lich-su-ban-giao": {
      "get": {
        "tags": [
          "2. Dịch vụ khai thác dữ liệu thông tin tài sản qua LGSP"
        ],
        "summary": "Lịch sử bàn giao tài sản",
        "operationId": "getLichSuBanGiao",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Số thứ tự trang (bắt đầu từ 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số lượng bản ghi trên một trang",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy dữ liệu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponsePagingBaseResponseBienBanBanGiaoDtoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "description": "API tra cứu lịch sử bàn giao vòng đời của một tài sản."
      }
    },
    "/v1/ban-giao": {
      "get": {
        "tags": [
          "2. Dịch vụ khai thác dữ liệu thông tin tài sản qua LGSP"
        ],
        "summary": "Tra cứu danh sách biên bản bàn giao",
        "operationId": "searchBienBan",
        "parameters": [
          {
            "name": "soBienBan",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tuNgay",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "denNgay",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "loaiBenNhan",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "CAN_BO",
                "DON_VI"
              ]
            }
          },
          {
            "name": "idBenNhan",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "loaiBenGiao",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "CAN_BO",
                "DON_VI",
                "KHO"
              ]
            }
          },
          {
            "name": "idBenGiao",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nguoiKyId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maDonVi",
            "in": "query",
            "description": "Lọc theo đơn vị liên quan (bên giao hoặc bên nhận)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trangThai",
            "in": "query",
            "description": "Lọc theo trạng thái biên bản (NHAP hoặc DA_KY)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Số thứ tự trang (bắt đầu từ 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số lượng bản ghi trên một trang",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy dữ liệu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponsePagingBaseResponseBienBanBanGiaoDtoResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequestError" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        },
        "description": "API tìm kiếm và lọc danh sách các biên bản bàn giao (phục vụ nghiệp vụ văn thư)."
      },
      "post": {
        "tags": [
          "2. Dịch vụ khai thác dữ liệu thông tin tài sản qua LGSP"
        ],
        "summary": "Lập biên bản bàn giao tài sản",
        "operationId": "banGiaoTaiSan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BienBanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lập biên bản thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponseBienBanBanGiaoDtoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "409": {
            "description": "Xung đột dữ liệu (Conflict). Xảy ra khi dữ liệu thêm mới bị trùng lặp mã số (ví dụ: so_bien_ban đã tồn tại).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "DUPLICATE_OR_CONSTRAINT_ERROR"
                    },
                    "message": {
                      "type": "string",
                      "example": "Dữ liệu đã tồn tại hoặc vi phạm ràng buộc cơ sở dữ liệu: Key (so_bien_ban)=(BB-NEW-001) already exists."
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "description": "API thực hiện nghiệp vụ bàn giao tài sản."
      }
    },
    "/v1/ban-giao/{id}": {
      "get": {
        "tags": [
          "2. Dịch vụ khai thác dữ liệu thông tin tài sản qua LGSP"
        ],
        "summary": "Xem chi tiết biên bản",
        "operationId": "getChiTietBienBan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponseBienBanBanGiaoDtoResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequestError" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/v1/ban-giao/{id}/ky": {
      "put": {
        "tags": [
          "2. Dịch vụ khai thác dữ liệu thông tin tài sản qua LGSP"
        ],
        "summary": "Phê duyệt biên bản",
        "operationId": "pheDuyetBienBan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PheDuyetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponseBienBanBanGiaoDtoResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequestError" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/v1/danh-muc/{loai}": {
      "get": {
        "tags": [
          "3. Dịch vụ chia sẻ dữ liệu danh mục qua LGSP"
        ],
        "summary": "Lấy dữ liệu danh mục",
        "operationId": "getDanhMuc",
        "parameters": [
          {
            "name": "loai",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "dia-gioi-hanh-chinh",
                "dan-toc",
                "ton-giao",
                "don-vi",
                "loai-tai-san"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Số thứ tự trang (bắt đầu từ 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số lượng bản ghi trên một trang",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy dữ liệu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainResponsePagingBaseResponseDanhMucResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "description": "API chia sẻ dữ liệu danh mục."
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequestError": {
        "description": "Lỗi dữ liệu đầu vào hoặc không tìm thấy",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BadRequestResponse"
            },
            "example": {
              "title": "Lỗi thao tác",
              "errorCode": "400",
              "message": "Dữ liệu đầu vào không hợp lệ hoặc không tìm thấy bản ghi tương ứng",
              "time": "2024-05-15T10:00:00"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Không tìm thấy dữ liệu",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BadRequestResponse"
            },
            "example": {
              "title": "Lỗi dữ liệu",
              "errorCode": "404",
              "message": "Không tìm thấy bản ghi tương ứng",
              "time": "2024-05-15T10:00:00"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Lỗi xác thực",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BadRequestResponse"
            },
            "example": {
              "title": "Lỗi xác thực",
              "errorCode": "401",
              "message": "Token không hợp lệ",
              "time": "2024-05-15T10:00:00"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Lỗi hệ thống máy chủ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BadRequestResponse"
            },
            "example": {
              "title": "Lỗi hệ thống",
              "errorCode": "500",
              "message": "Có lỗi xảy ra trong quá trình xử lý trên máy chủ",
              "time": "2024-05-15T10:00:00"
            }
          }
        }
      }
    },
    "schemas": {
...
      "PheDuyetRequest": {
        "type": "object",
        "required": [
          "nguoiKyId"
        ],
        "properties": {
          "nguoiKyId": {
            "type": "string",
            "example": "CB-04"
          }
        }
      },
...
      "MainResponseCanBoChiTietResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "Thành công"
          },
          "data": {
            "$ref": "#/components/schemas/CanBoChiTiet"
          }
        }
      },
      "MainResponseTaiSanDtoResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "Thành công"
          },
          "data": {
            "$ref": "#/components/schemas/TaiSanDto"
          }
        }
      },
      "MainResponseBienBanBanGiaoDtoResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "Thành công"
          },
          "data": {
            "$ref": "#/components/schemas/BienBanBanGiaoDto"
          }
        }
      },
      "MainResponsePagingBaseResponseCanBoTomTatResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "Thành công"
          },
          "data": {
            "type": "object",
            "properties": {
              "pageSize": {
                "type": "integer",
                "example": 10
              },
              "pageNumber": {
                "type": "integer",
                "example": 0
              },
              "totalPages": {
                "type": "integer",
                "example": 1
              },
              "totalElements": {
                "type": "integer",
                "example": 5
              },
              "content": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CanBoTomTat"
                }
              }
            }
          }
        }
      },
      "MainResponsePagingBaseResponseTaiSanDtoResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "Thành công"
          },
          "data": {
            "type": "object",
            "properties": {
              "pageSize": {
                "type": "integer",
                "example": 10
              },
              "pageNumber": {
                "type": "integer",
                "example": 0
              },
              "totalPages": {
                "type": "integer",
                "example": 1
              },
              "totalElements": {
                "type": "integer",
                "example": 5
              },
              "content": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TaiSanDto"
                }
              }
            }
          }
        }
      },
      "MainResponsePagingBaseResponseBienBanBanGiaoDtoResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "Thành công"
          },
          "data": {
            "type": "object",
            "properties": {
              "pageSize": {
                "type": "integer",
                "example": 10
              },
              "pageNumber": {
                "type": "integer",
                "example": 0
              },
              "totalPages": {
                "type": "integer",
                "example": 1
              },
              "totalElements": {
                "type": "integer",
                "example": 5
              },
              "content": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BienBanBanGiaoDto"
                }
              }
            }
          }
        }
      },
...
    }
  }
}